@@ -35,16 +35,16 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  dio: ^4.0.0-prev2
+  dio: ^5.4.3+1
   shared_preferences: ^2.1.1
   get_it: ^7.6.0
   provider: ^6.0.5
-  rive: ^0.9.1
-  awesome_dialog: ^3.0.2
+  rive: 0.11.1
+  awesome_dialog: 3.1.0
   flutter_advanced_drawer: ^1.3.5
   salomon_bottom_bar: ^3.3.2
   flutter_image_slideshow: ^0.1.6
-  image_picker: ^0.8.7+5
+  image_picker: ^1.1.2
 
   icons_plus: ^5.0.0
   heroicons_flutter: ^0.3.0
@@ -53,9 +53,9 @@ dependencies:
   cached_network_image: ^3.3.0
   datetime_picker_formfield_new: ^2.1.0
   localization: ^2.1.0
-  geocoding: ^2.0.5
-  google_maps_flutter: ^2.3.1
-  location: ^5.0.0
+  geocoding: 2.1.1
+  google_maps_flutter: ^2.6.0
+  location: 5.0.3
   url_launcher: ^6.1.12
   card_swiper : ^3.0.0
   animated_splash_screen: ^1.3.0
@@ -64,9 +64,12 @@ dependencies:
   clock: ^1.1.1
   get: ^4.6.6
   change_app_package_name: ^1.1.0
+  google_maps_flutter_ios: 2.5.0
+  geocoding_ios: 2.1.1
 
 
 dev_dependencies:
+  flutter_launcher_icons: "^0.13.1"
 
   flutter_test:
     sdk: flutter
@@ -76,7 +79,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -109,13 +112,19 @@ flutter:
     - family: nor
       fonts:
         - asset: assets/font/nor.ttf
-#         - asset: fonts/Schyler-Italic.ttf
-#           style: italic
-#     - family: Trajan Pro
-#       fonts:
-#         - asset: fonts/TrajanPro.ttf
-#         - asset: fonts/TrajanPro_Bold.ttf
-#           weight: 700
+  #         - asset: fonts/Schyler-Italic.ttf
+  #           style: italic
+  #     - family: Trajan Pro
+  #       fonts:
+  #         - asset: fonts/TrajanPro.ttf
+  #         - asset: fonts/TrajanPro_Bold.ttf
+  #           weight: 700
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
+flutter_launcher_icons:
+  android: "launcher_icon"
+  ios: true
+  remove_alpha_ios: true
+  image_path: "assets/images/logo.png"
+  min_sdk_android: 21 # android min sdk min:16, default 21
