@@ -1,12 +1,10 @@
 import 'package:awesome_dialog/awesome_dialog.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:keswaty/String_Extensions.dart';
 import 'package:keswaty/Widgets/DefaultButton.dart';
 import 'package:keswaty/Widgets/colors.dart';
 import 'package:keswaty/Widgets/langage.dart';
 import 'package:keswaty/controller/insert_sginup_provider.dart';
-import 'package:keswaty/data/ini_shard.dart';
 import 'package:keswaty/data/models/sginup_model.dart';
 import 'package:keswaty/main.dart';
 import 'package:keswaty/view/Login_Resiter/Login/login.dart';
@@ -19,7 +17,7 @@ import '../Forget_Password/ForgetPasswordWidgets/NoAccountText.dart';
 
 
 class SginUpPage extends StatefulWidget {
-  const SginUpPage({Key? key}) : super(key: key);
+  const SginUpPage({super.key});
 
   @override
   State<SginUpPage> createState() => _SginUpPageState();
@@ -52,21 +50,23 @@ class _SginUpPageState extends State<SginUpPage> {
   final List<String> errors =[];
 
   void addError({required String error}){
-    if(!error.contains(error))
+    if(!error.contains(error)) {
       setState(() {
         errors.add(error);
       });
+    }
   }
 
   void removError({required String error}){
-    if(!error.contains(error))
+    if(!error.contains(error)) {
       setState(() {
         errors.remove(error);
       });
+    }
   }
 //----------------------
 
-  Language _language = Language();
+  final Language _language = Language();
 
   @override
   Widget build(BuildContext context) {
@@ -76,14 +76,14 @@ class _SginUpPageState extends State<SginUpPage> {
       child: Scaffold(
         backgroundColor:
         Theme.of(context).brightness == Brightness.dark
-            ?  Color.fromRGBO(33, 37, 25, 1)
-            :  Color.fromRGBO(240, 242, 245, 1),
+            ?  const Color.fromRGBO(33, 37, 25, 1)
+            :  const Color.fromRGBO(240, 242, 245, 1),
 
         appBar: AppBar(
           backgroundColor:
           Theme.of(context).brightness == Brightness.dark
-              ?  Color.fromRGBO(33, 37, 25, 1)
-              :  Color.fromRGBO(240, 242, 245, 1),
+              ?  const Color.fromRGBO(33, 37, 25, 1)
+              :  const Color.fromRGBO(240, 242, 245, 1),
 
           // elevation: 0,
           // title:  Text(_language.tAddpost(),
@@ -113,13 +113,13 @@ class _SginUpPageState extends State<SginUpPage> {
               children: [
 
                 Padding(
-                  padding: EdgeInsets.symmetric(horizontal: 25),
+                  padding: const EdgeInsets.symmetric(horizontal: 25),
                   child: Column(
                     children: [
                       Column(
                         children: [
 
-                          SizedBox(height: 20),
+                          const SizedBox(height: 20),
 
                           Text(
                             _language.RegisterAccount(),
@@ -129,14 +129,14 @@ class _SginUpPageState extends State<SginUpPage> {
 
                           // SginUpForm(),
 
-                          SizedBox(height: 15),
+                          const SizedBox(height: 15),
 
                         ],
                       ),
 
-                      SizedBox(height: 20),
+                      const SizedBox(height: 20),
 
-                      Container(
+                      SizedBox(
                           width: screenWidth,
                           child: TextFormField(
                             onChanged: (value) {
@@ -151,13 +151,13 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration:  InputDecoration(
                                 labelText: _language.tFullname(),
                                 helperText: "",
-                                hintText: _language.tEnteryourname(),hintStyle: TextStyle(fontSize: 10),
-                                suffixIcon: Icon(
+                                hintText: _language.tEnteryourname(),hintStyle: const TextStyle(fontSize: 10),
+                                suffixIcon: const Icon(
                                     Icons.person_outline_sharp, color: kblack38)
                             ),
                           )
                       ),
-                      SizedBox(height: 10),
+                      const SizedBox(height: 10),
                       SizedBox(
                           width: screenWidth,
                           child: TextFormField(
@@ -175,14 +175,14 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration:  InputDecoration(
                                 labelText: _language.tPhonnumber(),
                                 helperText: "",
-                                hintText: _language.tenterphonenumber(),hintStyle: TextStyle(fontSize: 10),
-                                suffixIcon: Icon(Icons.phone, color: kblack38)
+                                hintText: _language.tenterphonenumber(),hintStyle: const TextStyle(fontSize: 10),
+                                suffixIcon: const Icon(Icons.phone, color: kblack38)
 
                             ),
                           )
                       ),
-                      SizedBox(height: 10),
-                      Container(
+                      const SizedBox(height: 10),
+                      SizedBox(
                           width: screenWidth,
                           child: TextFormField(
                             onChanged: (value) {
@@ -198,15 +198,15 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration:  InputDecoration(
                                 labelText: _language.temail(),
                                 helperText: "",
-                                hintText: _language.tEnterYourEmail(),hintStyle: TextStyle(fontSize: 10),
-                                suffixIcon: Icon(
+                                hintText: _language.tEnterYourEmail(),hintStyle: const TextStyle(fontSize: 10),
+                                suffixIcon: const Icon(
                                     Icons.email_outlined, color: kblack38)
                             ),
 
                           )
                       ),
-                      SizedBox(height: 10),
-                      Container(
+                      const SizedBox(height: 10),
+                      SizedBox(
                           width: screenWidth,
                           child: TextFormField(
                             onSaved: (newValue) => password = newValue!,
@@ -218,7 +218,7 @@ class _SginUpPageState extends State<SginUpPage> {
                               else if (value.length >= 8) {
                                 removError(error: KShortPassError);
                               }
-                              return null;
+                              return;
                             },
                             validator: (value) {
                               if (value!.isEmpty) {
@@ -238,7 +238,7 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration: InputDecoration(
                               labelText: _language.tpassword(),
                               helperText: "",
-                              hintText: _language.teEnteryourpassword(),hintStyle: TextStyle(fontSize: 10),
+                              hintText: _language.teEnteryourpassword(),hintStyle: const TextStyle(fontSize: 10),
 
                               suffixIcon: IconButton(
                                 onPressed: () =>
@@ -253,8 +253,8 @@ class _SginUpPageState extends State<SginUpPage> {
                             ),
                           )
                       ),
-                      SizedBox(height: 10),
-                      Container(
+                      const SizedBox(height: 10),
+                      SizedBox(
                           width: screenWidth,
                           child: TextFormField(
                             onChanged: (value) {
@@ -269,14 +269,14 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration:  InputDecoration(
                                 labelText: _language.taddress(),
                                 helperText: "",
-                                hintText: _language.teEnterAddress(),hintStyle: TextStyle(fontSize: 10),
-                                suffixIcon: Icon(
+                                hintText: _language.teEnterAddress(),hintStyle: const TextStyle(fontSize: 10),
+                                suffixIcon: const Icon(
                                     Icons.location_on_outlined, color: kblack38)
                             ),
                           )
                       ),
-                      SizedBox(height: 10),
-                      Container(
+                      const SizedBox(height: 10),
+                      SizedBox(
                           width: screenWidth,
                           child: TextFormField(
                             onChanged: (value) {
@@ -291,14 +291,14 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration:  InputDecoration(
                                 labelText: _language.tquestion(),
                                 helperText: "",
-                                hintText: _language.EntertheQuestion(),hintStyle: TextStyle(fontSize: 10),
-                                suffixIcon: Icon(
+                                hintText: _language.EntertheQuestion(),hintStyle: const TextStyle(fontSize: 10),
+                                suffixIcon: const Icon(
                                     Icons.location_on_outlined, color: kblack38)
                             ),
                           )
                       ),
-                      SizedBox(height: 10),
-                      Container(
+                      const SizedBox(height: 10),
+                      SizedBox(
                           width: screenWidth,
                           child: TextFormField(
                             onChanged: (value) {
@@ -313,13 +313,13 @@ class _SginUpPageState extends State<SginUpPage> {
                             decoration:  InputDecoration(
                                 labelText: _language.tanswer(),
                                 helperText: "",
-                                hintText: _language.tEnterAnswer(),hintStyle: TextStyle(fontSize: 10),
-                                suffixIcon: Icon(
+                                hintText: _language.tEnterAnswer(),hintStyle: const TextStyle(fontSize: 10),
+                                suffixIcon: const Icon(
                                     Icons.location_on_outlined, color: kblack38)
                             ),
                           )
                       ),
-                      SizedBox(height: 10),
+                      const SizedBox(height: 10),
 
 
                       // Consumer<InsertUserProvider>(builder: (context, value, child) {
@@ -328,7 +328,7 @@ class _SginUpPageState extends State<SginUpPage> {
                       Consumer<InsertSginUpProvider>(builder: (context, value, child) {
                         return
 
-                        Container(
+                        SizedBox(
 
 
                             width: screenWidth,
@@ -357,17 +357,17 @@ class _SginUpPageState extends State<SginUpPage> {
                                      Provider.of<InsertSginUpProvider>(context, listen: false).insert_response(sginUpModel);
 
                                     AwesomeDialog(
-                                      dialogBackgroundColor: Theme.of(context).brightness == Brightness.dark ? Color.fromRGBO(41, 45, 33, 1) : kwhait,
+                                      dialogBackgroundColor: Theme.of(context).brightness == Brightness.dark ? const Color.fromRGBO(41, 45, 33, 1) : kwhait,
                                       context: context,
                                       dialogType: DialogType.noHeader,
                                       animType: AnimType.topSlide,
                                       showCloseIcon: true,
                                       title: "",
                                       desc:_language.RegistrDon(),
-                                      descTextStyle: TextStyle(
+                                      descTextStyle: const TextStyle(
                                           fontWeight: FontWeight.bold),
                                       btnOkOnPress: () {
-                                        Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>Login()));
+                                        Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>const Login()));
                                       },
                                       btnOkColor: kblueColor,
                                       btnOkIcon: Icons.check_circle,
@@ -407,22 +407,22 @@ class _SginUpPageState extends State<SginUpPage> {
 
                               },
                               text: _language.Registration(),
-                              txtstyle: TextStyle(fontSize: 15, color: kwhait),
-                              icon: Icon(Icons.add, color: kwhait),)
+                              txtstyle: const TextStyle(fontSize: 15, color: kwhait),
+                              icon: const Icon(Icons.add, color: kwhait),)
                         );
                       }),
 
-                      SizedBox(height: 30,),
+                      const SizedBox(height: 30,),
 
 
 
-                      SizedBox(height: 10),
+                      const SizedBox(height: 10),
                       NoAccountText(
 
                         text1: _language.Bycontinuingconfirm(),
                         text2: "",
                       ),
-                      SizedBox(height: 40),
+                      const SizedBox(height: 40),
                     ],
                   ),
                 ),
