@@ -1,24 +1,23 @@
-import 'dart:convert';
-import 'dart:developer';
 
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:dio/dio.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:keswaty/Widgets/colors.dart';
 import 'package:keswaty/Widgets/langage.dart';
 import 'package:keswaty/main.dart';
 
+import '../../../constant.dart';
+
 
 class galleryus extends StatefulWidget {
-  const galleryus({Key? key}) : super(key: key);
+  const galleryus({super.key});
 
   @override
   _galleryusState createState() => _galleryusState();
 }
 
 class _galleryusState extends State<galleryus> {
-  Language _language = Language();
+  final Language _language = Language();
    static var storage;
   static var jsonList;
 
@@ -67,12 +66,12 @@ class _galleryusState extends State<galleryus> {
         backgroundColor: Theme
             .of(context)
             .brightness == Brightness.dark
-            ? Color.fromRGBO(33, 37, 25, 1)
-            : Color.fromRGBO(240, 242, 245, 1),
+            ? const Color.fromRGBO(33, 37, 25, 1)
+            : const Color.fromRGBO(240, 242, 245, 1),
           appBar: AppBar(
             backgroundColor: Theme.of(context).brightness == Brightness.dark
-                ?  Color.fromRGBO(33, 37, 25, 1)
-                :  Color.fromRGBO(240, 242, 245, 1),
+                ?  const Color.fromRGBO(33, 37, 25, 1)
+                :  const Color.fromRGBO(240, 242, 245, 1),
             elevation: 0,
             title:  Text(_language.Gallery(),
                 style: TextStyle(
@@ -99,20 +98,38 @@ class _galleryusState extends State<galleryus> {
 
                           children: [
                             Container(
+                              width: double.infinity,
+                              margin: const EdgeInsets.all(0.5),
                               child: Card(
                                 color: Theme
                                     .of(context)
                                     .brightness == Brightness.dark
-                                    ? Color.fromRGBO(41, 45, 33, 1)
+                                    ? const Color.fromRGBO(41, 45, 33, 1)
                                     : kwhait,
+                                elevation: 1,
+                                margin: const EdgeInsets.all(0.5),
 
 
                                 child: Column(
                                   children: [
-                                    Image.network('${_galleryusState.storage}${_galleryusState.jsonList[i]['PostPic'].toString()}', height: 250,
-                                        width: MediaQuery.of(context).size.width,
-                                        fit: BoxFit.cover),
-                                    Padding(padding: EdgeInsets.all(2),
+                                    CachedNetworkImage(
+                                      placeholder: (ctx, url) => Image.asset(AppConstants.image),
+                                      // fit: BoxFit.cover,
+                                      errorWidget: (ctx, url, err) =>
+                                          Image.asset(AppConstants.image,
+                                            // fit: BoxFit.cover,
+
+                                          ),
+                                      width: MediaQuery.of(context).size.width,
+                                      fit: BoxFit.cover,
+                                      height: 250,
+                                      imageUrl: '${_galleryusState.storage}${_galleryusState.jsonList[i]['PostPic'].toString()}',
+                                    ),
+                                    // Image.network('${_galleryusState.storage}${_galleryusState.jsonList[i]['PostPic'].toString()}', height: 250,
+                                    //     width: MediaQuery.of(context).size.width,
+                                    //
+                                    //     fit: BoxFit.cover),
+                                    Padding(padding: const EdgeInsets.all(2),
                                         child: Column(
                                           children: [
                                             Text(_galleryusState.jsonList[i]['PostTitle'].toString(),
@@ -121,17 +138,13 @@ class _galleryusState extends State<galleryus> {
                                                   fontSize: 15),
                                               textAlign: TextAlign.center,
                                             ),
-                                            SizedBox(height: 10,),
+                                            const SizedBox(height: 10,),
                                           ],
                                         )
                                     )
                                   ],
                                 ),
-                                elevation: 1,
-                                margin: const EdgeInsets.all(0.5),
                               ),
-                              width: double.infinity,
-                              margin: const EdgeInsets.all(0.5),
                             )
 
                           ],
