@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:awesome_dialog/awesome_dialog.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:icons_plus/icons_plus.dart';
 import 'package:keswaty/String_Extensions.dart';
@@ -20,9 +19,9 @@ class AddDonationPage extends StatefulWidget {
 
 
 
-  AddDonationPage({Key? key ,
+  const AddDonationPage({super.key ,
     // required this.bskt_id
-  }) : super(key: key);
+  });
 
 
   // DonationBsktModel bskt_id;
@@ -48,8 +47,9 @@ class _AddDonationPageState extends State<AddDonationPage> {
   final TextEditingController controllerLat = TextEditingController();
   final TextEditingController controllerLong = TextEditingController();
 
-  Language _language = Language();
+  final Language _language = Language();
 
+  @override
   void initState() {
     // TODO: implement initState
     super.initState();
@@ -69,12 +69,12 @@ class _AddDonationPageState extends State<AddDonationPage> {
               textDirection: language == 'AR' ? TextDirection.rtl : TextDirection.ltr,
               child: Scaffold(
                 backgroundColor: Theme.of(context).brightness == Brightness.dark
-                    ? Color.fromRGBO(33, 37, 25, 1)
-                    : Color.fromRGBO(240, 242, 245, 1),
+                    ? const Color.fromRGBO(33, 37, 25, 1)
+                    : const Color.fromRGBO(240, 242, 245, 1),
                 appBar: AppBar(
                   backgroundColor: Theme.of(context).brightness == Brightness.dark
-                      ? Color.fromRGBO(33, 37, 25, 1)
-                      : Color.fromRGBO(240, 242, 245, 1),
+                      ? const Color.fromRGBO(33, 37, 25, 1)
+                      : const Color.fromRGBO(240, 242, 245, 1),
                   elevation: 0,
                   leading: IconButton(
                     onPressed: () {
@@ -91,7 +91,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
                 body:
                 Consumer<DonationBsktGetProvider>(builder: (context, donationinsertPro, child) {
                   return donationinsertPro.isloading
-                      ? Center(
+                      ? const Center(
                           child: CircularProgressIndicator(),
                         )
                       :
@@ -105,8 +105,8 @@ class _AddDonationPageState extends State<AddDonationPage> {
                             child: Material(
                                 color: Theme.of(context).brightness ==
                                     Brightness.dark
-                                    ? Color.fromRGBO(41, 45, 33, 1)
-                                    : Color.fromRGBO(240, 242, 245, 1),
+                                    ? const Color.fromRGBO(41, 45, 33, 1)
+                                    : const Color.fromRGBO(240, 242, 245, 1),
                                 elevation: 0,
                                 borderRadius: BorderRadius.circular(5),
                                 clipBehavior: Clip.antiAliasWithSaveLayer,
@@ -115,28 +115,28 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                   children: [
                                     Column(
                                       children: [
-                                        SizedBox(
+                                        const SizedBox(
                                           height: 5,
                                         ),
                                         Container(
-                                          decoration: BoxDecoration(
+                                          decoration: const BoxDecoration(
                                             color: Colors.transparent,
                                             //  border: Border.all(color: kblueColor),
                                             borderRadius: BorderRadius.only(
                                                 topRight: Radius.circular(5),
                                                 topLeft: Radius.circular(5)),
                                           ),
-                                          child: Icon(
+                                          // Ink.image(image: AssetImage("images/found.png"),
+                                          height: 40,
+                                          width: screenWidth - 20,
+                                          child: const Icon(
                                             MingCute.hand_heart_line,
                                             color: greenColor,
                                             size: 40,
                                           ),
-                                          // Ink.image(image: AssetImage("images/found.png"),
-                                          height: 40,
-                                          width: screenWidth - 20,
                                           // fit: BoxFit.cover,
                                         ),
-                                        SizedBox(height: 6),
+                                        const SizedBox(height: 6),
 
                                         // SizedBox(height: 6,)
                                       ],
@@ -144,7 +144,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                   ],
                                 )),
                           ),
-                          SizedBox(height: 10),
+                          const SizedBox(height: 10),
                         ],
                       ),
                       Row(
@@ -157,10 +157,10 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                     : greenColor,
                               )),
                           Padding(
-                            padding: EdgeInsets.symmetric(horizontal: 10),
+                            padding: const EdgeInsets.symmetric(horizontal: 10),
                             child: Text(
                               _language.Donate(),
-                              style: TextStyle(
+                              style: const TextStyle(
                                   fontSize: 20,
                                   color: greenColor,
                                   fontWeight: FontWeight.bold),
@@ -175,7 +175,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
                               )),
                         ],
                       ),
-                      SizedBox(
+                      const SizedBox(
                         height: 15,
                       ),
                       //---------------------------------------------------------------------------------------------------------
@@ -183,10 +183,10 @@ class _AddDonationPageState extends State<AddDonationPage> {
                           key: _formKey,
                           child: Column(
                             children: [
-                              SizedBox(
+                              const SizedBox(
                                 height: 20,
                               ),
-                              Container(
+                              SizedBox(
                                 width: screenWidth - 20,
 
                                 child: DropdownButtonFormField<String>(
@@ -194,9 +194,10 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                     if (value == null) {
                                       return _language.tPleaseselectType();
                                     }
+                                    return null;
                                   },
                                   // value: lostfound_type,
-                                  items: donationinsertPro.donationtypedropdownlist!.map((DonationTypeModel element) {
+                                  items: donationinsertPro.donationtypedropdownlist.map((DonationTypeModel element) {
                                     return DropdownMenuItem(
                                       value: element.id.toString(),
                                       child: Text(element.name.toString()),
@@ -215,11 +216,11 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                 ),
                               ),
 
-                              SizedBox(height: 10),
+                              const SizedBox(height: 10),
 
-                              SizedBox(width: 20),
+                              const SizedBox(width: 20),
                               Text(_language.AddPhoto()),
-                              SizedBox(width: 10),
+                              const SizedBox(width: 10),
 
 
                               Row(
@@ -234,16 +235,16 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                           backgroundColor:  kblueColor,
                                           shape: RoundedRectangleBorder(
                                             borderRadius:BorderRadius.only(
-                                              bottomRight: language == 'AR' ? Radius.circular(15) : Radius.circular(0),
-                                              topRight: language == 'AR' ? Radius.circular(15) : Radius.circular(0),
-                                              bottomLeft: language == 'AR' ? Radius.circular(0) : Radius.circular(15),
-                                              topLeft: language == 'AR' ? Radius.circular(0) : Radius.circular(15),
+                                              bottomRight: language == 'AR' ? const Radius.circular(15) : const Radius.circular(0),
+                                              topRight: language == 'AR' ? const Radius.circular(15) : const Radius.circular(0),
+                                              bottomLeft: language == 'AR' ? const Radius.circular(0) : const Radius.circular(15),
+                                              topLeft: language == 'AR' ? const Radius.circular(0) : const Radius.circular(15),
                                             ),
                                           )
                                       ),
-                                      icon:Icon(Icons.camera_alt,color: Colors.white),
-                                      label: Text(_language.Camera(),style: TextStyle(color: Colors.white),)),
-                                  SizedBox(width: 1),
+                                      icon:const Icon(Icons.camera_alt,color: Colors.white),
+                                      label: Text(_language.Camera(),style: const TextStyle(color: Colors.white),)),
+                                  const SizedBox(width: 1),
                                   ElevatedButton.icon(onPressed:  () {
                                     donationinsertPro.pickMultipleImage();
                                   },
@@ -252,16 +253,16 @@ class _AddDonationPageState extends State<AddDonationPage> {
 
                                           shape: RoundedRectangleBorder(
                                             borderRadius:BorderRadius.only(
-                                              bottomRight: language == 'AR' ? Radius.circular(0) : Radius.circular(15),
-                                              topRight: language == 'AR' ? Radius.circular(0) : Radius.circular(15),
-                                              bottomLeft: language == 'AR' ? Radius.circular(15) : Radius.circular(0),
-                                              topLeft: language == 'AR' ? Radius.circular(15) : Radius.circular(0),
+                                              bottomRight: language == 'AR' ? const Radius.circular(0) : const Radius.circular(15),
+                                              topRight: language == 'AR' ? const Radius.circular(0) : const Radius.circular(15),
+                                              bottomLeft: language == 'AR' ? const Radius.circular(15) : const Radius.circular(0),
+                                              topLeft: language == 'AR' ? const Radius.circular(15) : const Radius.circular(0),
                                             ),
                                           )
                                       ),
-                                      icon:Icon(Icons.image,color: Colors.white,),
-                                      label: Text(_language.Gallery(),style: TextStyle(color: Colors.white))),
-                                  SizedBox(width: 10),
+                                      icon:const Icon(Icons.image,color: Colors.white,),
+                                      label: Text(_language.Gallery(),style: const TextStyle(color: Colors.white))),
+                                  const SizedBox(width: 10),
                                 ],
                               ),
 
@@ -270,7 +271,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                 shrinkWrap: true,
                                 itemCount: donationinsertPro.imageFileList?.length,
                                 padding: const EdgeInsets.all(20),
-                                gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+                                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
                                   crossAxisCount: 3,
                                   childAspectRatio: 1 / 1,
                                   mainAxisSpacing: 8,
@@ -303,11 +304,11 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                   );
                                 },
                               ),
-                              SizedBox(
+                              const SizedBox(
                                 height: 20,
                               ),
-                              SizedBox(height: 10),
-                              Container(
+                              const SizedBox(height: 10),
+                              SizedBox(
                                 width: screenWidth - 20,
                                 child:
                                 DefaultButton(
@@ -320,7 +321,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
 
 
 
-                                      if(donationinsertPro.imageFileList!.length > 0 ){
+                                      if(donationinsertPro.imageFileList!.isNotEmpty ){
                                         donationinsertPro.uploadImage(donationinsertPro.donationbsktadded_item.last.id.toString() ,id!);
 
 
@@ -335,7 +336,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                             btnOkText: _language.tbtnYse(),
                                             btnOkColor: kblueColor,
                                             btnOkIcon: Icons.check,
-                                            buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                                            buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                                             btnOkOnPress: (){
                                               Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_) => DonationsPage(donation: donationinsertPro.donationbsktadded_item.last)));
                                             }
@@ -352,7 +353,7 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                             btnOkText: _language.tbtnYse(),
                                             btnOkColor: kblueColor,
                                             btnOkIcon: Icons.check,
-                                            buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                                            buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                                             btnOkOnPress: (){
                                             }
                                         ).show();
@@ -375,8 +376,8 @@ class _AddDonationPageState extends State<AddDonationPage> {
                                   },
                                   text: _language.tadd(),
                                   txtstyle:
-                                  TextStyle(fontSize: 15, color: kwhait),
-                                  icon: Icon(Icons.add_circle_outline,
+                                  const TextStyle(fontSize: 15, color: kwhait),
+                                  icon: const Icon(Icons.add_circle_outline,
                                       color: Colors.white),
                                 ),
                               ),
@@ -388,27 +389,27 @@ class _AddDonationPageState extends State<AddDonationPage> {
                       //------------------------------------------------------------------------------------------
                       Column(
                         children: [
-                          SizedBox(height: 20),
+                          const SizedBox(height: 20),
                           Row(
                             children: [
-                              SizedBox(width: 20),
-                              Expanded(child: Divider()),
+                              const SizedBox(width: 20),
+                              const Expanded(child: Divider()),
                               Padding(
-                                padding: EdgeInsets.symmetric(horizontal: 10),
+                                padding: const EdgeInsets.symmetric(horizontal: 10),
                                 child: Text(
                                   _language.tNote(),
-                                  style: TextStyle(color: Colors.red),
+                                  style: const TextStyle(color: Colors.red),
                                 ),
                               ),
-                              Expanded(child: Divider()),
-                              SizedBox(width: 20),
+                              const Expanded(child: Divider()),
+                              const SizedBox(width: 20),
                             ],
                           ),
-                          SizedBox(height: 10),
+                          const SizedBox(height: 10),
                           Text(_language.dontnnote(),
                               textAlign: TextAlign.center,
                               style: TextStyleGray),
-                          SizedBox(height: 40),
+                          const SizedBox(height: 40),
                         ],
                       )
                     ],
