@@ -13,10 +13,9 @@ import 'package:keswaty/view/about_project/About_project.dart';
 import 'package:keswaty/view/about_project/con_us.dart';
 import 'package:keswaty/view/donation/add_baskets_page.dart';
 import 'package:keswaty/view/partners/partners_page.dart';
-import 'package:url_launcher/url_launcher.dart' as launcher;
 
 class Menus extends StatefulWidget {
-  const Menus({Key? key}) : super(key: key);
+  const Menus({super.key});
 
   @override
   _MenusState createState() => _MenusState();
@@ -24,7 +23,7 @@ class Menus extends StatefulWidget {
 
 class _MenusState extends State<Menus> {
 
-  Language _language = Language();
+  final Language _language = Language();
 
 
 
@@ -75,19 +74,19 @@ class _MenusState extends State<Menus> {
       Scaffold(
       backgroundColor:
       Theme.of(context).brightness == Brightness.dark
-          ?  Color.fromRGBO(33, 37, 25, 1)
-          :  Color.fromRGBO(240, 242, 245, 1),
+          ?  const Color.fromRGBO(33, 37, 25, 1)
+          :  const Color.fromRGBO(240, 242, 245, 1),
 
 
 
       body:ListView(
-       padding: EdgeInsets.all(8.0),
+       padding: const EdgeInsets.all(8.0),
         children: [
       Container(
         child: Container(
           margin: const EdgeInsets.symmetric(horizontal: 10,vertical: 10),
           // padding: EdgeInsets.only(bottom: 70),
-          decoration: BoxDecoration(
+          decoration: const BoxDecoration(
 
             borderRadius: BorderRadius.only(
               topRight: Radius.circular(70)
@@ -106,7 +105,7 @@ class _MenusState extends State<Menus> {
         //   crossAxisCount: 2,
         // crossAxisSpacing: 10, mainAxisSpacing: 1,
         shrinkWrap: true,
-        physics: NeverScrollableScrollPhysics(),
+        physics: const NeverScrollableScrollPhysics(),
         // childAspectRatio: 10,
         // mainAxisSpacing: 10,
 
@@ -124,10 +123,10 @@ class _MenusState extends State<Menus> {
                   btnOkText: _language.tbtnYse(),
                   btnOkColor: kblueColor,
 
-                  buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                  buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                   btnOkOnPress: (){
                     FocusScope.of(context).unfocus();
-                    Navigator.of(context).push(MaterialPageRoute(builder: (_)=>Login()));
+                    Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const Login()));
                   },
                   btnCancelOnPress: (){}
               ).show():
@@ -157,10 +156,10 @@ class _MenusState extends State<Menus> {
                               btnOkText: _language.tbtnYse(),
                               btnOkColor: kblueColor,
 
-                              buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                              buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                               btnOkOnPress: (){
                                 FocusScope.of(context).unfocus();
-                                Navigator.of(context).push(MaterialPageRoute(builder: (_)=>Login()));
+                                Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const Login()));
                               },
                               btnCancelOnPress: (){}
                           ).show():
@@ -173,7 +172,7 @@ class _MenusState extends State<Menus> {
                               Icon(MingCute.truck_line,size: screenWidth / 10, color:kwhait, ),
 
                             Text(
-                              _language.Createreceiptrequest(),style: TextStyle(fontSize:screenWidth / 25,color: kwhait),
+                              _language.Createreceiptrequest(),style: TextStyle(fontSize:screenWidth / 40,color: kwhait),
                             ),
 
 
@@ -196,7 +195,7 @@ class _MenusState extends State<Menus> {
                 desc: _language.tAlertSoon(),
                 btnOkText: _language.tbtnYse(),
                 btnOkColor: kblueColor,
-                buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                 btnOkOnPress: (){
                 },
               ).show();
@@ -204,7 +203,7 @@ class _MenusState extends State<Menus> {
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -223,7 +222,7 @@ class _MenusState extends State<Menus> {
                               desc: _language.tAlertSoon(),
                               btnOkText: _language.tbtnYse(),
                               btnOkColor: kblueColor,
-                              buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                              buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                               btnOkOnPress: (){
                               },
                           ).show();
@@ -248,11 +247,11 @@ class _MenusState extends State<Menus> {
           InkWell(
             onTap: (){
 
-              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>PartNersPAge(type_id: 1,)));
+              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const PartNersPAge(type_id: 1,)));
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -264,7 +263,7 @@ class _MenusState extends State<Menus> {
                         hoverColor: greenColor,
                         onTap: (){
 
-                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>PartNersPAge(type_id: 1,)));
+                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const PartNersPAge(type_id: 1,)));
                         },
                         child: Column(
                           children: [
@@ -289,11 +288,11 @@ class _MenusState extends State<Menus> {
           InkWell(
             onTap: () async {
 
-              await Navigator.of(context).push(MaterialPageRoute(builder: (_)=>PartNersPAge(type_id: 2)));
+              await Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const PartNersPAge(type_id: 2)));
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -306,7 +305,7 @@ class _MenusState extends State<Menus> {
 
                         onTap: () async {
 
-                          await Navigator.of(context).push(MaterialPageRoute(builder: (_)=>PartNersPAge(type_id: 2)));
+                          await Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const PartNersPAge(type_id: 2)));
                         },
                         child: Column(
                           children: [
@@ -331,11 +330,11 @@ class _MenusState extends State<Menus> {
           InkWell(
             onTap: (){
 
-              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>galleryus()));
+              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const galleryus()));
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -347,7 +346,7 @@ class _MenusState extends State<Menus> {
                         hoverColor: greenColor,
                         onTap: (){
 
-                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>galleryus()));
+                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const galleryus()));
                         },
                         child: Column(
                           children: [
@@ -415,12 +414,12 @@ class _MenusState extends State<Menus> {
           InkWell(
             onTap: (){
 
-              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>con_us()));
+              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const con_us()));
               // Navigator.pushNamed(context, "AddFound");
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -432,17 +431,17 @@ class _MenusState extends State<Menus> {
                         hoverColor: greenColor,
                         onTap: (){
 
-                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>con_us()));
+                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const con_us()));
                           // Navigator.pushNamed(context, "AddFound");
                         },
                         child: Column(
                           children: [
-                            SizedBox(height: 5,),
+                            const SizedBox(height: 5,),
 
                               Icon(FontAwesome.at_solid, color:kblueColor,size: screenWidth / 10, ),
 
                             Text(
-                              _language.socialmedia(),style: TextStyle(fontSize: screenWidth / 25,color: kblueColor),
+                              _language.socialmedia(),style: TextStyle(fontSize: screenWidth / 40,color: kblueColor),
                             ),
 
                           ],
@@ -459,11 +458,11 @@ class _MenusState extends State<Menus> {
           InkWell(
             onTap: (){
 
-              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>NewsPage()));
+              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const NewsPage()));
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -475,7 +474,7 @@ class _MenusState extends State<Menus> {
                         hoverColor: greenColor,
                         onTap: (){
 
-                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>NewsPage()));
+                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const NewsPage()));
                         },
                         child: Column(
                           children: [
@@ -499,11 +498,11 @@ class _MenusState extends State<Menus> {
           ),
           InkWell(
             onTap: (){
-              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>aboutpro()));
+              Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const aboutpro()));
             },
             child: Material(
                 color: Theme.of(context).brightness == Brightness.dark
-                    ?  Color.fromRGBO(41, 45, 33, 1)
+                    ?  const Color.fromRGBO(41, 45, 33, 1)
                     : kwhait,
                 elevation: 3,
                 borderRadius: BorderRadius.circular(15),
@@ -514,7 +513,7 @@ class _MenusState extends State<Menus> {
                     InkWell(
                         hoverColor: greenColor,
                         onTap: (){
-                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>aboutpro()));
+                          Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const aboutpro()));
                         },
                         child: Column(
                           children: [
@@ -525,7 +524,7 @@ class _MenusState extends State<Menus> {
                             Text(
                               _language.AboutProject(),style: TextStyle(fontSize:screenWidth / 25,color: kblueColor),
                             ),
-                            SizedBox(height: 6,)
+                            const SizedBox(height: 6,)
 
                           ],
                         )
@@ -548,7 +547,7 @@ class _MenusState extends State<Menus> {
     )
       ),
 
-          SizedBox(height: 30,)
+          const SizedBox(height: 30,)
         ],
       )
 
