@@ -8,9 +8,9 @@ import 'package:keswaty/main.dart';
 
 
 class PersonalData extends StatefulWidget {
-  PersonalData( {Key? key,
+  const PersonalData( {super.key,
 
-  }) : super(key: key);
+  });
 
 
   @override
@@ -18,7 +18,7 @@ class PersonalData extends StatefulWidget {
 }
 
 class _PersonalDataState extends State<PersonalData> {
-  Language _language = Language();
+  final Language _language = Language();
 
   @override
   Widget build(BuildContext context) {
@@ -29,12 +29,12 @@ class _PersonalDataState extends State<PersonalData> {
       child: Scaffold(
         backgroundColor:
         Theme.of(context).brightness == Brightness.dark
-            ?  Color.fromRGBO(33, 37, 25, 1)
-            :  Color.fromRGBO(240, 242, 245, 1),
+            ?  const Color.fromRGBO(33, 37, 25, 1)
+            :  const Color.fromRGBO(240, 242, 245, 1),
         appBar: AppBar(
           backgroundColor: Theme.of(context).brightness == Brightness.dark
-              ?  Color.fromRGBO(33, 37, 25, 1)
-              :  Color.fromRGBO(240, 242, 245, 1),
+              ?  const Color.fromRGBO(33, 37, 25, 1)
+              :  const Color.fromRGBO(240, 242, 245, 1),
           elevation: 0,
           title:  Text(_language.tProfile(),
               style: TextStyle(
@@ -54,7 +54,7 @@ class _PersonalDataState extends State<PersonalData> {
 
         body: Column(
           children: [
-            SizedBox(height: 20,),
+            const SizedBox(height: 20,),
 
             // Consumer<UserProvider>(builder: (context, userdprov, child) {
             //   return userdprov.isloading ? Center(child: CircularProgressIndicator(),) :
@@ -73,12 +73,12 @@ class _PersonalDataState extends State<PersonalData> {
                                       children: [
                                         Container(
                                           width: screenWidth,
-                                          padding: EdgeInsets.only(left: 10,right: 10,bottom: 10),
+                                          padding: const EdgeInsets.only(left: 10,right: 10,bottom: 10),
                                           child: Column(
                                             children: [
                                               Material(
                                                   color: Theme.of(context).brightness == Brightness.dark
-                                                      ? Color.fromRGBO(41, 45, 33, 1)
+                                                      ? const Color.fromRGBO(41, 45, 33, 1)
                                                       : kwhait,
                                                   elevation: 2,
                                                   borderRadius: BorderRadius.circular(5),
@@ -101,14 +101,14 @@ class _PersonalDataState extends State<PersonalData> {
                                                                       ? kblueColor
                                                                       : kblueColor,
                                                                 ),
-                                                                padding: EdgeInsets.only(left: 0,),
-                                                                margin: EdgeInsets.only(left: 0,top: 0),
+                                                                padding: const EdgeInsets.only(left: 0,),
+                                                                margin: const EdgeInsets.only(left: 0,top: 0),
 
 
                                                                 child: Row(
                                                                   mainAxisAlignment: MainAxisAlignment.start,
                                                                   children: [
-                                                                    SizedBox(width: 20,),
+                                                                    const SizedBox(width: 20,),
                                                                     Container(
                                                                       width: 60.0,
                                                                       height: 60.0,
@@ -139,11 +139,11 @@ class _PersonalDataState extends State<PersonalData> {
                                                                                 : kwhait,
                                                                         ),
                                                                       ),),
-                                                                    SizedBox(width: 15),
+                                                                    const SizedBox(width: 15),
                                                                     Column(
                                                                       children: [
                                                                               if(InitSharedPreferences.getNameUser()!=null&&InitSharedPreferences.getNameUser()!.isNotEmpty)
-                                                                                Container(
+                                                                                SizedBox(
                                                                                   width: screenWidth / 2,
                                                                                   child: Text(InitSharedPreferences.getNameUser()!, style: TextStyle(fontWeight: FontWeight.bold,fontSize: screenWidth / 30,
                                                                                     color: Theme.of(context).brightness == Brightness.dark
@@ -157,7 +157,7 @@ class _PersonalDataState extends State<PersonalData> {
 
                                                                       ],
                                                                     ),
-                                                                    SizedBox(width: 10),
+                                                                    const SizedBox(width: 10),
                                                                   ],
                                                                 ) ,
                                                               ),
@@ -167,109 +167,109 @@ class _PersonalDataState extends State<PersonalData> {
 
                                                       ),
                                                       Container(
-                                                        decoration:  BoxDecoration(
+                                                        decoration:  const BoxDecoration(
                                                           // border: Border(
                                                           //   left:BorderSide(color: kblueColor,width: 5) ,),
                                                         ),
                                                         child: Column(
                                                           children: [
-                                                            SizedBox(height: 15,),
+                                                            const SizedBox(height: 15,),
                                                             Column(
                                                               mainAxisAlignment: MainAxisAlignment.start,
                                                               children: [
                                                                 Row(
                                                                   children: [
-                                                                    SizedBox(width: 15,),
-                                                                    Icon(Icons.person_outline_sharp,size: 20,),
-                                                                    SizedBox(width: 5),
+                                                                    const SizedBox(width: 15,),
+                                                                    const Icon(Icons.person_outline_sharp,size: 20,),
+                                                                    const SizedBox(width: 5),
                                                                     Text(_language.tNamee()),
                                                                   ],
                                                                 ),
                                                                 const SizedBox(height: 10),
                                                                 Row(children: [
-                                                                  SizedBox(width: 40),
+                                                                  const SizedBox(width: 40),
 
                                                                   if(InitSharedPreferences.getNameUser()!=null&&InitSharedPreferences.getNameUser()!.isNotEmpty)
                                                                     Text(
-                                                                      InitSharedPreferences.getNameUser()!, style: TextStyle(fontSize: 13,color: kblack38),
+                                                                      InitSharedPreferences.getNameUser()!, style: const TextStyle(fontSize: 13,color: kblack38),
                                                                     
                                                                     )
                                                                 ],),
                                                               ],
                                                             ),
-                                                            SizedBox(height: 7,),
+                                                            const SizedBox(height: 7,),
                                                             buildDivider(),
-                                                            SizedBox(height: 7,),
+                                                            const SizedBox(height: 7,),
                                                             Column(
                                                               mainAxisAlignment: MainAxisAlignment.start,
                                                               children: [
                                                                 Row(
                                                                   children: [
-                                                                    SizedBox(width: 15,),
-                                                                    Icon(Icons.phone_outlined,size: 20,),
-                                                                    SizedBox(width: 5),
+                                                                    const SizedBox(width: 15,),
+                                                                    const Icon(Icons.phone_outlined,size: 20,),
+                                                                    const SizedBox(width: 5),
                                                                     Text(_language.tPhonee()),
                                                                   ],
                                                                 ),
-                                                                SizedBox(height: 10),
+                                                                const SizedBox(height: 10),
                                                                 Row(children: [
-                                                                  SizedBox(width: 40),
+                                                                  const SizedBox(width: 40),
                                                                   if(InitSharedPreferences.getPhoneUser()!=null&&InitSharedPreferences.getPhoneUser()!.isNotEmpty)
-                                                                    Text(InitSharedPreferences.getPhoneUser().toString(), style: TextStyle(fontSize: 13,color: kblack38),)
+                                                                    Text(InitSharedPreferences.getPhoneUser().toString(), style: const TextStyle(fontSize: 13,color: kblack38),)
                                                                 ],),
                                                               ],
                                                             ),
-                                                            SizedBox(height: 7,),
+                                                            const SizedBox(height: 7,),
                                                             buildDivider(),
-                                                            SizedBox(height: 7,),
+                                                            const SizedBox(height: 7,),
                                                             Column(
                                                               mainAxisAlignment: MainAxisAlignment.start,
                                                               children: [
                                                                 Row(
                                                                   children: [
-                                                                    SizedBox(width: 15,),
-                                                                    Icon(Icons.email_outlined,size: 20,),
-                                                                    SizedBox(width: 5),
+                                                                    const SizedBox(width: 15,),
+                                                                    const Icon(Icons.email_outlined,size: 20,),
+                                                                    const SizedBox(width: 5),
                                                                     Text(_language.tEmaill()),
                                                                   ],
                                                                 ),
-                                                                SizedBox(height: 10),
+                                                                const SizedBox(height: 10),
                                                                 Row(children: [
-                                                                  SizedBox(width: 40),
+                                                                  const SizedBox(width: 40),
                                                                   if(InitSharedPreferences.getEmailUser()!=null&&InitSharedPreferences.getEmailUser()!.isNotEmpty)
-                                                                    Text(InitSharedPreferences.getEmailUser()!, style: TextStyle(fontSize: 13,color: kblack38),)
+                                                                    Text(InitSharedPreferences.getEmailUser()!, style: const TextStyle(fontSize: 13,color: kblack38),)
                                                                   //Text("hassan@gmail.com",style: TextStyle(fontSize: 13, color: kblack38)),
                                                                 ],),
                                                               ],
                                                             ),
                                                         
-                                                            SizedBox(height: 7,),
+                                                            const SizedBox(height: 7,),
                                                             buildDivider(),
-                                                            SizedBox(height: 7,),
+                                                            const SizedBox(height: 7,),
 
                                                             Column(
                                                               mainAxisAlignment: MainAxisAlignment.start,
                                                               children: [
                                                                 Row(
                                                                   children: [
-                                                                    SizedBox(width: 15,),
-                                                                    Icon(Icons.location_on_outlined,size: 20,),
-                                                                    SizedBox(width: 5),
+                                                                    const SizedBox(width: 15,),
+                                                                    const Icon(Icons.location_on_outlined,size: 20,),
+                                                                    const SizedBox(width: 5),
                                                                     Text(_language.tadress()),
                                                                   ],
                                                                 ),
-                                                                SizedBox(height: 10),
+                                                                const SizedBox(height: 10),
                                                                 Row(children: [
-                                                                  SizedBox(width: 40),
+                                                                  const SizedBox(width: 40),
                                                                   if(InitSharedPreferences.getAddressUser()!=null&&InitSharedPreferences.getAddressUser()!.isNotEmpty)
-                                                                    Text(InitSharedPreferences.getAddressUser()!, style: TextStyle(fontSize: 13,color: kblack38))
+                                                                    Text(InitSharedPreferences.getAddressUser()!, style: const TextStyle(fontSize: 13,color: kblack38))
                                                                   //Text("sana'a al jaraf",style: TextStyle(fontSize: 13,color: kblack38)),
                                                                 ],),
                                                               ],
                                                             ),
 
 
-                                                            SizedBox(height: 15,),
+                                                            const SizedBox(height: 15,),
 
 
                                                           ],
@@ -286,7 +286,7 @@ class _PersonalDataState extends State<PersonalData> {
                               
 
 
-                                        SizedBox(height: 10),
+                                        const SizedBox(height: 10),
 
                                       ]
                                   ),
