@@ -1,18 +1,14 @@
-import 'package:awesome_dialog/awesome_dialog.dart';
     import 'package:flutter/material.dart';
     import 'package:keswaty/Widgets/colors.dart';
     import 'package:keswaty/Widgets/langage.dart';
-    import 'package:keswaty/data/ini_shard.dart';
     import 'package:keswaty/main.dart';
-    import 'package:keswaty/view/Login_Resiter/Login/login.dart';
-    import 'package:keswaty/view/setting/personal_data/personal_data.dart';
     import 'package:keswaty/view/setting/setting_widgets/setting_page_body.dart';
 
 
 
 
     class SettingPage extends StatefulWidget {
-    const SettingPage({Key? key}) : super(key: key);
+    const SettingPage({super.key});
 
     @override
     State<SettingPage> createState() => _SettingPageState();
@@ -20,7 +16,7 @@ import 'package:awesome_dialog/awesome_dialog.dart';
 
     class _SettingPageState extends State<SettingPage> {
 
-    Language _language = Language();
+    final Language _language = Language();
 
     @override
     Widget build(BuildContext context) {
@@ -30,12 +26,12 @@ import 'package:awesome_dialog/awesome_dialog.dart';
     child: Scaffold(
     backgroundColor:
     Theme.of(context).brightness == Brightness.dark
-    ?  Color.fromRGBO(33, 37, 25, 1)
-        :  Color.fromRGBO(240, 242, 245, 1),
+    ?  const Color.fromRGBO(33, 37, 25, 1)
+        :  const Color.fromRGBO(240, 242, 245, 1),
     appBar: AppBar(
     backgroundColor: Theme.of(context).brightness == Brightness.dark
-    ?  Color.fromRGBO(33, 37, 25, 1)
-        :  Color.fromRGBO(240, 242, 245, 1),
+    ?  const Color.fromRGBO(33, 37, 25, 1)
+        :  const Color.fromRGBO(240, 242, 245, 1),
     elevation: 0,
     title:  Text(_language.tSettings(),
               style: TextStyle(
@@ -56,7 +52,7 @@ import 'package:awesome_dialog/awesome_dialog.dart';
           ),
         ),
          body: ListView(
-             children: [
+             children: const [
 
                SettingPageBody(),
 
