@@ -1,6 +1,5 @@
 
 import 'package:animated_splash_screen/animated_splash_screen.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:keswaty/view/navigationbar/nav.dart';
 
@@ -13,18 +12,18 @@ class splsh extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
 
-    return MaterialApp(
+    return const MaterialApp(
       // debugShowCheckedModeBanner: false,
       // theme: ThemeData(
       //   useMaterial3: true,
       // ),
-      home: const mysplsh(),
+      home: mysplsh(),
     );
   }
 }
 
 class mysplsh extends StatefulWidget {
-  const mysplsh({Key? key}) : super(key: key);
+  const mysplsh({super.key});
 
   @override
   State<mysplsh> createState() => _mysplshState();
@@ -40,7 +39,7 @@ class _mysplshState extends State<mysplsh> {
           .of(context)
           .brightness == Brightness.dark
           ? Image.asset('assets/images/splshblack.gif'):Image.asset('assets/images/splshwhite.gif'),
-        nextScreen: nav(),
+        nextScreen: const nav(),
         duration: 2300,
         backgroundColor:  Theme
             .of(context)
