@@ -1,17 +1,5 @@
 
-
-import 'package:flutter/material.dart';
-import 'package:keswaty/Widgets/My_Input_Them.dart';
-
-GetMaterialApp(
-
-
-ThemeData(
-
-
-
-)
-) {
+GetMaterialApp(Function() ThemeData) {
   // TODO: implement GetMaterialApp
   throw UnimplementedError();
 }
