@@ -1,5 +1,4 @@
 import 'package:awesome_dialog/awesome_dialog.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_advanced_drawer/flutter_advanced_drawer.dart';
 import 'package:icons_plus/icons_plus.dart';
@@ -22,7 +21,7 @@ import 'package:salomon_bottom_bar/salomon_bottom_bar.dart';
 
 
 class nav extends StatefulWidget {
-  nav({super.key,});
+  const nav({super.key,});
 
   @override
   State<nav> createState() => _navState();
@@ -34,7 +33,7 @@ class _navState extends State<nav> {
 
   String page = 'Home';
 
-  Language _language = Language();
+  final Language _language = Language();
 
   static var _currentIndex = 0;
 
@@ -50,7 +49,7 @@ class _navState extends State<nav> {
       width: 10,
           decoration: BoxDecoration(
               color: Theme.of(context).brightness == Brightness.dark
-                  ?  Color.fromRGBO(33, 37, 25, 1)
+                  ?  const Color.fromRGBO(33, 37, 25, 1)
                   :  kwhait,
 
           ),
@@ -65,7 +64,7 @@ class _navState extends State<nav> {
         disabledGestures: false,
 
         childDecoration: const BoxDecoration(
-          borderRadius: const BorderRadius.all(Radius.circular(16)),
+          borderRadius: BorderRadius.all(Radius.circular(16)),
         ),
 
         drawer: Directionality(
@@ -81,19 +80,19 @@ class _navState extends State<nav> {
                           child: Column(
 
                             children: [
-                              SizedBox(height: 60),
+                              const SizedBox(height: 60),
                               Container(
                                 child: InkWell(
                                   onTap: () {
                                     if (InitSharedPreferences.getNameUser() != null &&
-                                        InitSharedPreferences.getNameUser()!.isNotEmpty)
-                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => PersonalData()));
-                                    else
+                                        InitSharedPreferences.getNameUser()!.isNotEmpty) {
+                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => const PersonalData()));
+                                    } else {
                                       AwesomeDialog(
                                           dialogBackgroundColor:
                                           Theme.of(context).brightness ==
                                               Brightness.dark
-                                              ? Color.fromRGBO(41, 45, 33, 1)
+                                              ? const Color.fromRGBO(41, 45, 33, 1)
                                               : kwhait,
                                           context: context,
                                           dialogType: DialogType.noHeader,
@@ -106,19 +105,20 @@ class _navState extends State<nav> {
                                           btnOkColor: kblueColor,
                                           btnOkIcon: Icons.login,
                                           buttonsBorderRadius:
-                                          BorderRadius.all(Radius.circular(10)),
+                                          const BorderRadius.all(Radius.circular(10)),
                                           btnOkOnPress: () {
-                                            Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_) => Login()));
+                                            Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_) => const Login()));
                                           },
                                           btnCancelOnPress: () {})
                                           .show();
+                                    }
                                   },
                                   child: Center(
                                       child: Column(
                                         children: [
                                           Container(
-                                            padding: EdgeInsets.only(left: 0,),
-                                            margin: EdgeInsets.only(left: 0, top: 0),
+                                            padding: const EdgeInsets.only(left: 0,),
+                                            margin: const EdgeInsets.only(left: 0, top: 0),
                                             child: Row(
                                               children: [
 
@@ -172,7 +172,7 @@ class _navState extends State<nav> {
                                                         children: [
                                                           if (InitSharedPreferences.getNameUser() != null &&
                                                               InitSharedPreferences.getNameUser()!.isNotEmpty)
-                                                            Container(
+                                                            SizedBox(
                                                               width: screenWidth / 2,
                                                               child: Text(InitSharedPreferences.getNameUser()!,
                                                                   style: TextStyle(fontWeight: FontWeight.bold, fontSize: screenWidth / 30, color:
@@ -206,7 +206,7 @@ class _navState extends State<nav> {
 
                                                         if (InitSharedPreferences.getEmailUser() != null &&
                                                             InitSharedPreferences.getEmailUser()!.isNotEmpty)
-                                                          Container(
+                                                          SizedBox(
                                                             width: screenWidth / 2,
                                                             child: Text(
                                                               overflow: TextOverflow.ellipsis,
@@ -246,15 +246,15 @@ class _navState extends State<nav> {
                               ListTile(
                                     onTap: () {
                                       if (InitSharedPreferences.getNameUser() != null &&
-                                          InitSharedPreferences.getNameUser()!.isNotEmpty)
+                                          InitSharedPreferences.getNameUser()!.isNotEmpty) {
                                         Navigator.of(context).push(
-                                            MaterialPageRoute(builder: (_) => FeedbackPage()));
-                                      else {
+                                            MaterialPageRoute(builder: (_) => const FeedbackPage()));
+                                      } else {
                                         AwesomeDialog(
                                             dialogBackgroundColor:
                                             Theme.of(context).brightness ==
                                                 Brightness.dark
-                                                ? Color.fromRGBO(41, 45, 33, 1)
+                                                ? const Color.fromRGBO(41, 45, 33, 1)
                                                 : kwhait,
                                             context: context,
                                             dialogType: DialogType.noHeader,
@@ -267,11 +267,11 @@ class _navState extends State<nav> {
                                             btnOkColor: kblueColor,
                                             btnOkIcon: Icons.login,
                                             buttonsBorderRadius:
-                                            BorderRadius.all(Radius.circular(10)),
+                                            const BorderRadius.all(Radius.circular(10)),
                                             btnOkOnPress: () {
                                               Navigator.of(context).pushReplacement(
                                                   MaterialPageRoute(
-                                                      builder: (_) => SginUpPage()));
+                                                      builder: (_) => const SginUpPage()));
                                             },
                                             btnCancelOnPress: () {})
                                             .show();
@@ -289,7 +289,7 @@ class _navState extends State<nav> {
                                   ),
                               ListTile(
                                     onTap: () {
-                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => SettingPage()));
+                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => const SettingPage()));
                                     },
                                     leading: Icon(EvaIcons.settings_outline,
                                         color: Theme.of(context).brightness == Brightness.dark
@@ -304,7 +304,7 @@ class _navState extends State<nav> {
                               ListTile(
                                     onTap: () {
 
-                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => HelpPage()));
+                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => const HelpPage()));
                                     },
                                     leading: Icon(Icons.question_mark,
                                         color: Theme.of(context).brightness == Brightness.dark
@@ -319,7 +319,7 @@ class _navState extends State<nav> {
                                   ),
                               ListTile(
                                     onTap: () {
-                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => aboutpro()));
+                                      Navigator.of(context).push(MaterialPageRoute(builder: (_) => const aboutpro()));
                                       },
 
                                     leading: Icon(Icons.info_outline,
@@ -341,7 +341,7 @@ class _navState extends State<nav> {
                                           dialogBackgroundColor:
                                           Theme.of(context).brightness ==
                                               Brightness.dark
-                                              ? Color.fromRGBO(41, 45, 33, 1)
+                                              ? const Color.fromRGBO(41, 45, 33, 1)
                                               : kwhait,
                                           context: context,
                                           dialogType: DialogType.question,
@@ -352,7 +352,7 @@ class _navState extends State<nav> {
                                           btnOkColor: kblueColor,
                                           btnOkIcon: Icons.check,
                                           buttonsBorderRadius:
-                                          BorderRadius.all(Radius.circular(10)),
+                                          const BorderRadius.all(Radius.circular(10)),
                                           btnOkText: _language.tbtnYse(),
                                           btnOkOnPress: () async {
                                             await InitSharedPreferences.logOut();
@@ -360,7 +360,7 @@ class _navState extends State<nav> {
                                               dialogBackgroundColor:
                                               Theme.of(context).brightness ==
                                                   Brightness.dark
-                                                  ? Color.fromRGBO(41, 45, 33, 1)
+                                                  ? const Color.fromRGBO(41, 45, 33, 1)
                                                   : kwhait,
                                               context: context,
                                               dialogType: DialogType.noHeader,
@@ -371,7 +371,7 @@ class _navState extends State<nav> {
                                               btnOkColor: kblueColor,
                                               btnOkIcon: Icons.check,
                                               buttonsBorderRadius:
-                                              BorderRadius.all(Radius.circular(10)),
+                                              const BorderRadius.all(Radius.circular(10)),
                                               btnOkOnPress: () async {
                                                 await InitSharedPreferences.logOut();
                                                 setState(() {});
@@ -383,7 +383,7 @@ class _navState extends State<nav> {
                                           btnCancelText: _language.tbtnCancel())
                                           .show();
                                     } else {
-                                      Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>Login()));
+                                      Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>const Login()));
                                     }
                                   },
                                   leading: Icon(
@@ -398,7 +398,7 @@ class _navState extends State<nav> {
                                           ? _language.tlogin()
                                           : _language.tlogout()
                                       ,
-                                      style: TextStyle(color: kblueColor,fontWeight: FontWeight.bold))
+                                      style: const TextStyle(color: kblueColor,fontWeight: FontWeight.bold))
                               ),
 
 
@@ -408,10 +408,10 @@ class _navState extends State<nav> {
 
                         Column(
                           children: [
-                            Text(_language.trights(),textAlign: TextAlign.center,style: TextStyle(
+                            Text(_language.trights(),textAlign: TextAlign.center,style: const TextStyle(
                                 color: kblack38,fontSize: 12
                             )),
-                            SizedBox(height: 40),
+                            const SizedBox(height: 40),
                           ],
                         ),
                       ],
@@ -426,8 +426,8 @@ class _navState extends State<nav> {
 
           appBar: AppBar(
             backgroundColor: Theme.of(context).brightness == Brightness.dark
-                ? Color.fromRGBO(33, 37, 25, 1)
-                : Color.fromRGBO(240, 242, 245, 1),
+                ? const Color.fromRGBO(33, 37, 25, 1)
+                : const Color.fromRGBO(240, 242, 245, 1),
             elevation: 0,
             title: Center(
               child:
@@ -450,7 +450,7 @@ class _navState extends State<nav> {
                 valueListenable: _advancedDrawerController,
                 builder: (_, value, __) {
                   return AnimatedSwitcher(
-                    duration: Duration(milliseconds: 250),
+                    duration: const Duration(milliseconds: 250),
                     child: Icon(
                       color: Theme.of(context).brightness == Brightness.dark
                           ? kwhait
@@ -462,7 +462,7 @@ class _navState extends State<nav> {
                 },
               ),
             ):
-                SizedBox(),
+                const SizedBox(),
 
 
             actions: [
@@ -482,10 +482,10 @@ class _navState extends State<nav> {
                       btnOkText: _language.tbtnYse(),
                       btnOkColor: kblueColor,
 
-                      buttonsBorderRadius: BorderRadius.all(Radius.circular(10)),
+                      buttonsBorderRadius: const BorderRadius.all(Radius.circular(10)),
                       btnOkOnPress: (){
                         FocusScope.of(context).unfocus();
-                        Navigator.of(context).push(MaterialPageRoute(builder: (_)=>Login()));
+                        Navigator.of(context).push(MaterialPageRoute(builder: (_)=>const Login()));
                       },
                     btnCancelOnPress: (){}
                   ).show():
@@ -505,7 +505,7 @@ class _navState extends State<nav> {
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: [
-                      SizedBox(width: 5,),
+                      const SizedBox(width: 5,),
                       Text(_currentIndex == 2?
                       _language.Createreceiptrequest() : "",
                         style: TextStyle(
@@ -513,12 +513,12 @@ class _navState extends State<nav> {
                             fontWeight: FontWeight.bold,fontSize: screenWidth / 30
                         ),
                       ),
-                      SizedBox(width: 5,),
+                      const SizedBox(width: 5,),
                     ],
                   ),
                 ),
               ):
-                  SizedBox()
+                  const SizedBox()
             ],
           ),
           // _pages[_currentIndex],
@@ -526,8 +526,8 @@ class _navState extends State<nav> {
 
           bottomNavigationBar: SalomonBottomBar(
             backgroundColor: Theme.of(context).brightness == Brightness.dark
-                ? Color.fromRGBO(33, 37, 25, 1)
-                : Color.fromRGBO(240, 242, 245, 1),
+                ? const Color.fromRGBO(33, 37, 25, 1)
+                : const Color.fromRGBO(240, 242, 245, 1),
 
 
             selectedItemColor: kblueColor,
@@ -588,11 +588,11 @@ class _navState extends State<nav> {
             builder: (context) {
               switch (_currentIndex) {
                 case 0:
-                  return Container(child: MainPage());
+                  return Container(child: const MainPage());
                 case 1:
                   if(InitSharedPreferences.getNameUser()!=null){
                     return Container(child: AddBasketsPage());
-                  } return Container(child: Login());
+                  } return Container(child: const Login());
 
                 case 2:
                   return Container(child: donation());
@@ -600,7 +600,7 @@ class _navState extends State<nav> {
 
                 case 3:
                    // if(InitSharedPreferences.getNameUser()!=null)
-                  return Container(child: Menus());
+                  return Container(child: const Menus());
 
                 default:
                   return Container();
